Remove changes for -Gec and fix unit test failure

+++ tools/clang/test/HLSLFileCheck/hlsl/semantics/legacy/obsolete_semantics_ps_3.hlsl
@@ -1,8 +1,8 @@
-// RUN: %dxc -E main -T ps_6_0 %s -Gec | FileCheck %s
+// RUN: %dxc -E main -T ps_6_0 %s | FileCheck %s
 
 // CHECK: warning: Ignoring unsupported 'VFACE' in the target attribute string
 
-float4 main(float4 color : COLOR, float vface : VFACE) : SV_TARGET
+float4 main(float4 color : COLOR, uint vface : VFACE) : SV_TARGET
 {
   return (vface > 0) ? color : (color*2);
 }
